Create DeskStop-Lite.py
This is the simple version of DeskStop. Upon running the program all the files on your desktop will be moved into a folder in your documents titled after the date you ran the clean.

This version has no whitelist capabilities and you cannot choose what directories to sort or where to sort them.